Move zod utils into core
