+++ src/core/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './path.js';
 export * from './service-initializer.js';
 export * from './api-key-resolver.js';
+export * from './zod-schema-converter.js';
